content: transformContent: move replacerSpec into #replacer
